# api-spec.yaml
openapi: 3.0.3
info:
  title: AgentHub Registry API
  description: |
    The universal package registry for AI-native agents, tools, chains, and prompts.

    ## Authentication

    This API uses GitHub OAuth 2.0 for authentication. Most endpoints require a valid JWT token.

    ## Rate Limiting

    API requests are rate limited to 1000 requests per minute per IP address.

    ## Package Types

    The registry supports the following package types:
    - `agent`: AI agents and assistants
    - `tool`: Utilities and functions for agents
    - `chain`: Workflow chains and pipelines
    - `prompt`: Prompt templates and examples
    - `dataset`: Training and reference datasets

    ## Versioning

    All packages follow semantic versioning (semver) conventions.
  version: 1.0.0
  contact:
    name: AgentHub Team
    url: https://github.com/agenthubcli/registry
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://registry.agenthubcli.com
    description: Production server
  - url: http://localhost:8000
    description: Development server
paths:
...
  /api/v1/search/:
    get:
      summary: Search packages
      description: Search for packages by name, description, or keywords
      tags: [Search]
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
            minLength: 1
            maxLength: 500
        - $ref: "#/components/parameters/PackageTypeFilter"
        - $ref: "#/components/parameters/SortBy"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"
        "422":
          description: Invalid search parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/search/popular:
    get:
      summary: Get popular packages
      description: Get packages sorted by download count
      tags: [Search]
      parameters:
        - $ref: "#/components/parameters/PackageTypeFilter"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Popular packages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"
  /api/v1/search/recent:
    get:
      summary: Get recent packages
      description: Get recently published packages
      tags: [Search]
      parameters:
        - $ref: "#/components/parameters/PackageTypeFilter"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Recent packages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"
  /api/v1/search/trending:
    get:
      summary: Get trending packages
      description: Get packages with high recent download activity
      tags: [Search]
      parameters:
        - $ref: "#/components/parameters/PackageTypeFilter"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Trending packages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"

  # User Endpoints
  /api/v1/users/{username}:
    get:
      summary: Get user profile
      description: Get public profile information for a user
      tags: [Users]
      parameters:
        - name: username
          in: path
          required: true
          description: GitHub username
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/users/{username}/packages:
    get:
      summary: Get user packages
      description: Get packages owned by a specific user
      tags: [Users]
      parameters:
        - name: username
          in: path
          required: true
          description: GitHub username
          schema:
            type: string
        - $ref: "#/components/parameters/PackageTypeFilter"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: User packages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPackages"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Health Check Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained via GitHub OAuth
  parameters:
...
    PackageTypeFilter:
      name: package_type
      in: query
      description: Filter by package type
      schema:
        type: string
        enum: [agent, tool, chain, prompt, dataset]
    SortBy:
      name: sort_by
      in: query
      description: Sort results by field
      schema:
        type: string
        enum: [relevance, downloads, created, updated]
        default: relevance
    Limit:
      name: limit
      in: query
      description: Maximum number of results
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
...
    PackageType:
      type: string
      enum: [agent, tool, chain, prompt, dataset]
      description: Type of package

    # Common Responses
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Error message
        error_code:
          type: string
          description: Machine-readable error code
      example:
        detail: "Package not found"
        error_code: "PACKAGE_NOT_FOUND"
...
    UserProfile:
      type: object
      required: [id, github_username]
      properties:
        id:
          type: integer
          example: 123
        github_username:
          type: string
          example: "johndoe"
        display_name:
          type: string
          example: "John Doe"
        github_avatar_url:
          type: string
          format: uri
          example: "https://github.com/avatars/johndoe"
        bio:
          type: string
          example: "AI researcher and developer"
        website:
          type: string
          format: uri
          example: "https://johndoe.dev"
        location:
          type: string
          example: "San Francisco, CA"
        company:
          type: string
          example: "AI Company Inc."
        total_packages:
          type: integer
          example: 5
        total_downloads:
          type: integer
          example: 1234
        created_at:
          type: string
          format: date-time

    # Package Upload
...
    Package:
      type: object
      required: [id, name, package_type, owner, created_at]
      properties:
        id:
          type: integer
          example: 456
        name:
          type: string
          example: "awesome-tool"
        description:
          type: string
          example: "An awesome tool for AI development"
        package_type:
          $ref: "#/components/schemas/PackageType"
        latest_version:
          type: string
          example: "1.2.3"
        total_downloads:
          type: integer
          example: 5678
        download_count_last_30_days:
          type: integer
          example: 234
        version_count:
          type: integer
          example: 8
        homepage:
          type: string
          format: uri
          example: "https://github.com/user/awesome-tool"
        repository:
          type: string
          format: uri
          example: "https://github.com/user/awesome-tool"
        documentation:
          type: string
          format: uri
          example: "https://awesome-tool.readthedocs.io"
        keywords:
          type: array
          items:
            type: string
          example: ["ai", "tool", "automation"]
        owner:
          $ref: "#/components/schemas/UserProfile"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SearchResults:
      type: object
      required: [results, total, limit, offset, query]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/Package"
        total:
          type: integer
          description: Total number of matching packages
          example: 156
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
        query:
          type: string
          example: "machine learning"
        package_type:
          $ref: "#/components/schemas/PackageType"
        sort_by:
          type: string
          example: "relevance"
    UserPackages:
      type: object
      required: [username, packages, total_packages]
      properties:
        username:
          type: string
          example: "johndoe"
        packages:
          type: array
          items:
            $ref: "#/components/schemas/Package"
        total_packages:
          type: integer
          example: 5
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
...